added unit test and integration test
